requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
idna==3.4
psycopg2-binary==2.9.6
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
